Bugfix update and delete service, refactored code.

+++ SpringProject/src/main/java/com/example/projectmrsisa/controller/RetreatController.java
@@ -5,6 +5,7 @@ import com.example.projectmrsisa.dto.RetreatDTO;
 import com.example.projectmrsisa.dto.ServiceAvailabilityDTO;
 import com.example.projectmrsisa.model.*;
 import com.example.projectmrsisa.service.*;
+import com.example.projectmrsisa.validators.Validator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -52,6 +53,8 @@ public class RetreatController {
     @Autowired
     private SubscriptionService subscriptionService;
 
+    private Validator validator = new Validator();
+
     @GetMapping(value="/getAll", produces = "application/json")
     public ResponseEntity<List<RetreatDTO>> getRetreats() {
         try {
@@ -74,8 +77,8 @@ public class RetreatController {
         try {
             retreatOwner = userService.findUserByEmail(loggedUser.getName());
         } catch (Exception e) { return new ResponseEntity<>(HttpStatus.BAD_REQUEST); }
-        if (!validAddress(retreatDTO)) return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
-        if (!validateRetreatData(retreatDTO)) return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        if (!validator.validAddress(retreatDTO.getCountry(), retreatDTO.getCity(), retreatDTO.getStreet())) return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        if (!validator.validateRetreatData(retreatDTO)) return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         Address a;
         Set<Tag> additionalServices;
         try {
@@ -84,40 +87,6 @@ public class RetreatController {
             Retreat retreat = retreatService.addRetreat(new Retreat(retreatDTO, a, additionalServices, retreatOwner));
             return new ResponseEntity<>(new RetreatDTO(retreat), HttpStatus.OK);
         }catch (Exception e) { return new ResponseEntity<>(HttpStatus.BAD_REQUEST); }
-    }
-
-    private boolean validateRetreatData(RetreatDTO retreatDTO) {
-        if (retreatDTO.getName() == null || retreatDTO.getName().equals("")){
-            return false;
-        }
-        if (retreatDTO.getDescription() == null || retreatDTO.getDescription().equals("")){
-            return false;
-        }
-        if (retreatDTO.getNumOfRooms() <= 0) {
-            return false;
-        }
-        for (String as: retreatDTO.getAdditionalServices()) {
-            if (as.equals("") || as.length() > 14) return false;
-        }
-        for (String rc: retreatDTO.getRulesOfConduct()) {
-            if (rc.equals("")) return false;
-            if (!rc.equals("Inside smoking disallowed") && !rc.equals("Pet friendly") && !rc.equals("Parties disallowed") && !rc.equals("Furniture braking is charged extra")) return false;
-        }
-        for (String pic: retreatDTO.getPictures()) {
-            if (pic.equals("")) return false;
-            if (!pic.endsWith(".png") && !pic.endsWith(".jpg") && !pic.endsWith(".jpeg")) return false;
-        }
-        return retreatDTO.getNumOfBeds() > 0;
-    }
-
-    private boolean validAddress(RetreatDTO retreatDTO) {
-        if (retreatDTO.getCountry().equals("") || retreatDTO.getCountry() == null || !retreatDTO.getCountry().matches("([A-Z]{1})([a-z]+)([^0-9]*)$")) {
-            return false;
-        }
-        if (retreatDTO.getCity().equals("") || retreatDTO.getCity() == null || !retreatDTO.getCity().matches("([A-Z]{1})([a-z]+)([^0-9]*)$")) {
-            return false;
-        }
-        return !retreatDTO.getStreet().equals("") && retreatDTO.getStreet() != null;
     }
 
     @GetMapping(value = "/get/{id}", produces = "application/json")
@@ -131,13 +100,12 @@ public class RetreatController {
         }
     }
 
-    @Transactional
     @PutMapping(value = "/update-retreat/{id}", produces = "application/json")
     @PreAuthorize("hasRole('retreatOwner')")
     public ResponseEntity<RetreatDTO> updateRetreat(@PathVariable Integer id, @RequestBody RetreatDTO retreatDTO) {
-        if (!validAddress(retreatDTO)) return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
-        if (!validateRetreatData(retreatDTO)) return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
-        // TODO: provera da li postoje rezervacije za vikendicu
+        if (!validator.validAddress(retreatDTO.getCountry(), retreatDTO.getCity(), retreatDTO.getStreet())) return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        if (!validator.validateRetreatData(retreatDTO)) return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        if (!reservationService.pendingReservationForServiceExists(id)) return new ResponseEntity<>(HttpStatus.CONFLICT);
         try {
             Retreat retreat = retreatService.getRetreatById(id);
             Set<Tag> newAdditionalServices = tagService.findTags(retreatDTO.getAdditionalServices(), "retreat");
@@ -148,13 +116,12 @@ public class RetreatController {
         }
     }
 
-    @Transactional
     @DeleteMapping(value = "/delete-retreat/{id}")
     @PreAuthorize("hasRole('retreatOwner')")
     public ResponseEntity<RetreatDTO> deleteRetreat(@PathVariable Integer id) {
         try {
-            // TODO: provera da li postoje rezervacije za vikendicu
             Retreat retreat = retreatService.getRetreatById(id);
+            if (!reservationService.pendingReservationForServiceExists(id)) return new ResponseEntity<>(HttpStatus.CONFLICT);
             if (retreat == null) return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
             retreatService.deleteRetreat(id);
             return new ResponseEntity<>(new RetreatDTO(retreat), HttpStatus.OK);
@@ -166,7 +133,7 @@ public class RetreatController {
     @PostMapping(value = "/add-availability/{id}")
     @PreAuthorize("hasRole('retreatOwner')")
     public ResponseEntity<ServiceAvailabilityDTO> addRetreatAvailability(@PathVariable Integer id, @RequestBody ServiceAvailabilityDTO serviceAvailabilityDTO) {
-        if (!validServiceAvailability(serviceAvailabilityDTO)) return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        if (!validator.validServiceAvailability(serviceAvailabilityDTO)) return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         try {
             Retreat retreat = retreatService.getRetreatById(id);
             if (retreat == null) return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
@@ -182,7 +149,7 @@ public class RetreatController {
     @PostMapping(value = "/add-action/{id}")
     @PreAuthorize("hasRole('retreatOwner')")
     public ResponseEntity<ActionDTO> addAction(@PathVariable Integer id, @RequestBody ActionDTO actionDTO, Principal loggedUser) {
-        if (!validAction(actionDTO)) return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        if (!validator.validAction(actionDTO)) return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         try {
             Retreat retreat = retreatService.getRetreatById(id);
             if (retreat == null) return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
@@ -199,34 +166,6 @@ public class RetreatController {
         }catch (Exception e) {
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
-    }
-
-    private boolean validServiceAvailability(ServiceAvailabilityDTO serviceAvailabilityDTO) {
-        if (!validDates(serviceAvailabilityDTO.getDateFrom(), serviceAvailabilityDTO.getDateTo(), serviceAvailabilityDTO.getTimeFrom(), serviceAvailabilityDTO.getTimeTo())) return false;
-        return true;
-    }
-    
-    private boolean validAction(ActionDTO actionDTO) {
-        if (!validDates(actionDTO.getDateFrom(), actionDTO.getDateTo(), actionDTO.getTimeFrom(), actionDTO.getTimeTo())) return false;
-        System.out.println(actionDTO.getMaxNumOfPeople());
-        if (actionDTO.getMaxNumOfPeople() <= 0) return false;
-        for (String as: actionDTO.getAdditionalServices()) {
-            if (as.equals("") || as.length() > 14) return false;
-        }
-        return !(actionDTO.getPrice() <= 0);
-    }
-
-    private boolean validDates(Date dateFrom, Date dateTo, String timeFrom, String timeTo) {
-        Date today = new Date();
-        if (dateFrom == null || dateTo == null || timeFrom == null || timeTo == null) return false;
-        if (dateFrom.compareTo(today) < 0) return false;
-        if (dateTo.compareTo(today) < 0 ) return false;
-        if (dateFrom.compareTo(dateTo) > 0) return false;
-        if (dateFrom.compareTo(dateTo) == 0) {
-            return Integer.parseInt(timeFrom) * 60 + Integer.parseInt(timeFrom)
-                    < Integer.parseInt(timeTo) * 60 + Integer.parseInt(timeTo);
-        }
-        return true;
     }
 
     @GetMapping(value = "/get-actions/{id}")
